+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+#### 26 June 2018
+
+- Add full CRUD for exchange rates under a new /exchangerates endpoint.
+
 #### 23 June 2018
 
 - Change README.md API documentation to point to http://andresamayadiaz.github.io/FrontAccountingSimpleAPI/